# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Account {
  id: number;
  name: string;
  institution: string;
  kind: string;
  balance: number;
  color: string;
}

export interface AccountInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  institution: string;
  /** @minLength 1 */
  kind: string;
  balance: number;
  /** @minLength 1 */
  color: string;
}

export interface Category {
  id: number;
  name: string;
  color: string;
  icon: string;
  spent: number;
  budget: number;
}

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  income: 'income',
  expense: 'expense',
  transfer: 'transfer',
} as const;

export interface Transaction {
  id: number;
  description: string;
  amount: number;
  type: TransactionType;
  category: string;
  account: string;
  date: string;
  status: string;
}

export type TransactionInputType = typeof TransactionInputType[keyof typeof TransactionInputType];


export const TransactionInputType = {
  income: 'income',
  expense: 'expense',
  transfer: 'transfer',
} as const;

export interface TransactionInput {
  /** @minLength 1 */
  description: string;
  /** @exclusiveMinimum 0 */
  amount: number;
  type: TransactionInputType;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  account: string;
  /** @minLength 1 */
  date: string;
}

export interface Goal {
  id: number;
  name: string;
  target: number;
  current: number;
  deadline: string;
  color: string;
}

export interface GoalInput {
  /** @minLength 1 */
  name: string;
  /** @exclusiveMinimum 0 */
  target: number;
  /** @minimum 0 */
  current: number;
  /** @minLength 1 */
  deadline: string;
  /** @minLength 1 */
  color: string;
}

export interface BudgetLine {
  category: string;
  planned: number;
  spent: number;
  color: string;
}

export interface Budget {
  month: string;
  planned: number;
  spent: number;
  lines: BudgetLine[];
}

export interface BudgetInput {
  /** @minLength 1 */
  month: string;
  /** @minimum 0 */
  planned: number;
  lines: BudgetLine[];
}

export interface CreditCard {
  id: number;
  name: string;
  institution: string;
  limit: number;
  used: number;
  closingDay: number;
  dueDay: number;
  color: string;
}

export type DashboardSummaryCashFlowItem = {
  month: string;
  income: number;
  expenses: number;
};

export interface DashboardSummary {
  netWorth: number;
  available: number;
  income: number;
  expenses: number;
  investments: number;
  bills: number;
  savingsRate: number;
  healthScore: number;
  healthLabel: string;
  monthlyChange: number;
  cashFlow: DashboardSummaryCashFlowItem[];
}

export type ListTransactionsParams = {
type?: ListTransactionsType;
search?: string;
/**
 * @maximum 100
 */
limit?: number;
};

export type ListTransactionsType = typeof ListTransactionsType[keyof typeof ListTransactionsType];


export const ListTransactionsType = {
  income: 'income',
  expense: 'expense',
  transfer: 'transfer',
} as const;