# lib/api-zod/src/generated/types/account.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Account {
  id: number;
  name: string;
  institution: string;
  kind: string;
  balance: number;
  color: string;
}