# lib/api-zod/src/generated/types/accountInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AccountInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  institution: string;
  /** @minLength 1 */
  kind: string;
  balance: number;
  /** @minLength 1 */
  color: string;
}