# lib/api-zod/src/generated/types/budget.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BudgetLine } from './budgetLine';

export interface Budget {
  month: string;
  planned: number;
  spent: number;
  lines: BudgetLine[];
}