# lib/api-zod/src/generated/types/budgetInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BudgetLine } from './budgetLine';

export interface BudgetInput {
  /** @minLength 1 */
  month: string;
  /** @minimum 0 */
  planned: number;
  lines: BudgetLine[];
}