# lib/api-zod/src/generated/types/budgetLine.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BudgetLine {
  category: string;
  planned: number;
  spent: number;
  color: string;
}