# lib/api-zod/src/generated/types/creditCard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreditCard {
  id: number;
  name: string;
  institution: string;
  limit: number;
  used: number;
  closingDay: number;
  dueDay: number;
  color: string;
}