# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryCashFlowItem } from './dashboardSummaryCashFlowItem';

export interface DashboardSummary {
  netWorth: number;
  available: number;
  income: number;
  expenses: number;
  investments: number;
  bills: number;
  savingsRate: number;
  healthScore: number;
  healthLabel: string;
  monthlyChange: number;
  cashFlow: DashboardSummaryCashFlowItem[];
}