# lib/api-zod/src/generated/types/goal.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Goal {
  id: number;
  name: string;
  target: number;
  current: number;
  deadline: string;
  color: string;
}