# lib/api-zod/src/generated/types/goalInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GoalInput {
  /** @minLength 1 */
  name: string;
  /** @exclusiveMinimum 0 */
  target: number;
  /** @minimum 0 */
  current: number;
  /** @minLength 1 */
  deadline: string;
  /** @minLength 1 */
  color: string;
}