# lib/api-zod/src/generated/types/transactionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TransactionInputType } from './transactionInputType';

export interface TransactionInput {
  /** @minLength 1 */
  description: string;
  /** @exclusiveMinimum 0 */
  amount: number;
  type: TransactionInputType;
  /** @minLength 1 */
  category: string;
  /** @minLength 1 */
  account: string;
  /** @minLength 1 */
  date: string;
}